requirements.txt
Django==3.0.*
djangorestframework
coreapi
Markdown
drf-yasg
ipython
